Refactor: Dted.java
Comments in english for Dted.java

+++ src/main/java/br/edu/inteli/cc/m5/grupo2/Dted.java
@@ -9,32 +9,29 @@ public class Dted {
 
     public static double[][] readDted(String filePath, int interval) {
 
-        // Abrir o arquivo DTED
         gdal.AllRegister();
         Dataset dataset = gdal.Open(filePath, gdalconst.GA_ReadOnly);
-
-        // Obter o objeto Band que contém os dados de altitude
         Band altitudeBand = dataset.GetRasterBand(1);
 
-        // Obter as dimensões da matriz de altitude
+        // Dimensions X and Y of the file
         int cols = altitudeBand.getXSize();
         int rows = altitudeBand.getYSize();
 
-        // Cria a matriz para armazenar os valores de altitude, latitude e longitude
+        // Creation of a matrix to hold the values of altitude, latitude, and longitude for each vertex
         double[][] data = new double[(rows * cols) / (interval / 30)][3];
 
-        // Obter os dados de latitude e longitude usando as funções GetGeoTransform
+        // Configuration variables of the file
         double[] geotransform = dataset.GetGeoTransform();
         double xOrigin = geotransform[0];
         double yOrigin = geotransform[3];
         double pixelWidth = geotransform[1];
         double pixelHeight = geotransform[5];
 
-        // Ler os valores de altitude em cada pixel usando a função ReadRaster
+        // Read the values of altitude in each pixel
         double[] buffer = new double[cols * rows];
         altitudeBand.ReadRaster(0, 0, cols, rows, buffer);
 
-        // Preencher a matriz com os valores de altitude, latitude e longitude
+        // Fill the matrix com os valores de altitude, latitude e longitude
         for (int i = 0; i < rows - interval / 30; i++) {
             for (int j = 0; j < cols - interval / 30; j++) {
                 double altitude = buffer[i * cols + j];
@@ -44,12 +41,12 @@ public class Dted {
                 data[index][0] = altitude;
                 data[index][1] = latitude;
                 data[index][2] = longitude;
-                j += (int) interval / 30;
+                j += interval / 30;
             }
-            i += (int) interval / 30;
+            i += interval / 30;
         }
 
-        // Fechar o objeto Dataset
+        // Close the Dataset object
         dataset.delete();
         return data;
     }
